wip save pages to shared cache
